requirements.txt
aiogram==3.6.0
python-dotenv==1.0.1
aiosqlite==0.20.0
setuptools>=70.0.0
